requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-pinecone>=0.0.1
langchain-huggingface>=0.0.1
openai>=1.3.0
pinecone>=2.2.4
sentence-transformers>=2.2.2
spacy>=3.7.0
pdfplumber>=0.10.0
pypdf>=3.15.1
python-dotenv>=1.0.0
numpy>=1.21.0
tqdm>=4.62.0
fastapi>=0.95.0
python-multipart>=0.0.6
uvicorn>=0.21.1
jinja2>=3.1.2
aiofiles>=23.1.0